applyed addOnce, FIRST and ONCE logic

+++ srcNew/consequence.js
@@ -1,10 +1,21 @@
 // @flow
 import ruleDB from './ruleDB'
-import type {Rule,Action,Store} from './types'
+import type {Rule,Action,Store,RuleContext} from './types'
 
 let store = null
 
-export default function consequence (rule:Rule, action:Action, store:Store, addRule:Function, removeRule:Function):boolean{
+export default function consequence (context:RuleContext, action:Action, store:Store, addRule:Function, removeRule:Function):boolean{
+  const rule = context.rule
+  // skip when concurrency matches
+  if(rule.concurrency === 'ONCE' && context.running){
+    return false
+  }
+  if(rule.concurrency === 'FIRST' && context.running){
+    return false
+  }
+  if(rule.addOnce && context.running){
+    return false
+  }
   // skip if 'skipRule' condition matched
   if(action.meta && action.meta.skipRule){
     const skipRules = Array.isArray(action.meta.skipRule) 
@@ -19,24 +30,29 @@ export default function consequence (rule:Rule, action:Action, store:Store, addR
     return false
   }
 
-  const result = rule.consequence(store, action, {addRule, removeRule})
+  context.running++
+  const result = rule.consequence(store, action)
 
   if(typeof result === 'object' && result.type){
     const action:any = result
     store.dispatch(action)
+    rule.concurrency !== 'ONCE' && context.running--
+    rule.addOnce && ruleDB.removeRule(rule)
   }
 
   else if(typeof result === 'object' && result.then){
     const promise:any = result
-    promise.then(action => action && action.type && store.dispatch(action)) 
+    promise.then(action => {
+      action && action.type && store.dispatch(action)
+      rule.concurrency !== 'ONCE' && context.running--
+      rule.addOnce && ruleDB.removeRule(rule)
+    }) 
   }
 
   else if(typeof result === 'function'){
     ruleDB.addUnlistenCallback(rule, () => result(store.getState))
-  }
-
-  if(rule.addOnce){
-    ruleDB.removeRule(rule)
+    rule.concurrency !== 'ONCE' && context.running--
+    rule.addOnce && ruleDB.removeRule(rule)
   }
 
   return true
